Use string table and flat array for transcript.
Previously, the output format was closer to display format with the
sentences grouped by speaker. This was to save some amount of disk
space. However, it makes future implementation of edits to the
transcript more difficult as the structure encodes inferrences about
relations between the data as well as display preferences.

Going back to a flat denormalized form ends up being a bit more
flexible even if it is more verbose and redundant.

Also start extracting out a "common" directory for code shared between
the node.js cloud functions envrionment and the rendering environment.