Convert Cloud functions to common/transcript
Redo the storage access layer to unify the code in both the React components
and the cloud functions.

Also fix up a huge number of typescript build config issues.

This this, build, watch, and test in the top-level as well as in funcitons
all work now.

Cloud function tests are run with firebase emulator:exec to fake out the
library dependencies.