Fix functions build failure with esm.
Making this work with

  * firebase functions
  * esm modules, esp with "type":"modules" in package.json
  * typescript

is hilariously hard.  Firebase functions do not understand esm modules. This
causes their preprocessor to blow up if you use them. There is a workaround
here:

  https://github.com/firebase/firebase-tools/issues/2994#issuecomment-1961379942

that uses esbuild to bundle all of the code into one huge index.js that
papers over all the preprocessing issues. It also does some magic with
package.json and the npm-run-all package to ensure all of esbuild, tsc,
and firebase run in parallel so that changing a typescript source file
results in an actual behavior change to a running function emulator.

However, this is still not sufficient because esbuild has decided to
bundle everything in devDependencies and dependencies into the index.js.
This will pull in old commonJS packages but attempt to run that
syntax in an ESM environment which causes a cryptic

  "Dynamic require of 'something' is not supported"

This will send you down rabbit holes of trying to fix tsc config because
it looks like you are exporting the wrong module loading type..which
will lead you to try setting tsconfig's module to "nodenext"...which
confusing *requires* you to put .js after all your relative path imports
even if the source file is .ts.  This is confusing as heck.

But the real problem is you need to exclude packages from esbuild:

  https://github.com/evanw/esbuild/issues/3324#issuecomment-1685217503

And now everything will run. Howeever, it means you have to ensure the
deploy picks up your node_modules.... ugh.

Oh...and jest's setup requires you to NOT use verbatimModuleSyntax otherwise
you cannot import typescript types since its config file is a CommonJs
setup and the import syntax is ESM.